+++ scripts/gc_test.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+def main():
+    print("test script")
+
+
+if __name__ == '__main__':
+    main()